EvoNN-Stratograph/pyproject.toml
[project]
name = "stratograph"
version = "0.1.0"
description = "Hierarchy-first evolutionary neural architecture search"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.1.3",
    "numpy>=2.2.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "typer>=0.15.1",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]
data = ["openml>=0.14", "pandas>=2.2"]
all = ["stratograph[dev,data]"]

[project.scripts]
stratograph = "stratograph.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stratograph"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["stratograph"]

[tool.coverage.report]
show_missing = true
skip_covered = true
